# backend/public/swagger.yaml
swagger: '2.0'

info:
  version: "0.0.1"
  title: Bank Customers

consumes:
  - text/plain

produces:
  - application/json

paths:
  /customers:
    get:
      tags:
        - customers
      operationId: getAll
      description: Get all customers
      responses:
        200:
          $ref: '#/responses/getAllCustomers'
    post:
      tags:
        - customers
      operationId: add
      description: Add new Customer
      parameters:
        - $ref: '#/parameters/dateParam'
        - $ref: '#/parameters/birthParam'
        - $ref: '#/parameters/incomeParam'
      responses:
        200:
          $ref: '#/responses/getOneCustomer'

  /customer/{id}:
    get:
      tags:
        - customers
      operationId: get
      description: Get customer by ID
      parameters:
        - $ref: '#/parameters/idParam'
      responses:
        200:
          $ref: '#/responses/getOneCustomer'



    delete:
      tags:
        - customers
      operationId: delete
      description: Delete customer by ID
      parameters:
        - $ref: '#/parameters/idParam'
      responses:
        200:
          $ref: '#/responses/getOneCustomer'


  /customer/{id}/goals:
    get:
      tags:
        - goals
      operationId: get
      description: Get goals using customer ID
      parameters:
        - $ref: '#/parameters/idParam'
      responses:
        200:
          $ref: '#/responses/getGoalsCustomer'

    post:
      tags:
        - goals
      operationId: add
      description: Add new Goal using customer id
      parameters:
        - $ref: '#/parameters/idParam'
        - $ref: '#/parameters/nameGoalParam'
        - $ref: '#/parameters/valueGoalParam'
        - $ref: '#/parameters/SavedGoalParam'
        - $ref: '#/parameters/dateGoalParam'
        - $ref: '#/parameters/accountGoalParam'
        - $ref: '#/parameters/priorityGoalParam'


      responses:
        200:
          $ref: '#/responses/getGoalsCustomer'


    delete:
      tags:
        - goals
      operationId: delete
      description: Delete a goal using customer id and name
      parameters:
        - $ref: '#/parameters/idParam'
        - $ref: '#/parameters/nameGoalParam'
      responses:
          $ref: '#/responses/getGoalsCustomer'



definitions:

  goal:
    type: object
    description: A goal object
    required:
      - value
      - name
      - date
      - priority
      - accountid

    properties:
      value:
        type: integer
      saved:
        type: integer
      name:
        type: string
      date:
        type: string
      priority:
        type: integer
      accountid:
        type: string

  customer:
    type: object
    description: A customer object
    required:
      - _id
      - _rev
      - birth
      - registrationdate
      - expectedincome
    properties:
      _id:
        type: integer
        format: int64
      _rev:
        type: integer
        format: int64
      birth:
        type: string
      registrationdate:
        type: string
      expectedincome:
        type: integer
      goals:
        type: array
        items:
          $ref: '#/definitions/goal'


responses:
  getOneCustomer:
    description: One customer
    schema:
      $ref: '#/definitions/customer'

  getAllCustomers:
    description: List of all customers
    schema:
      type: array
      items:
        $ref: '#/definitions/customer'

  getGoalsCustomer:
    description: Get the goals of one customer
    schema:
      type: array
      $ref: '#/definitions/goal'

parameters:
  idParam:
    name: id
    in: path
    description: Customer ID
    required: true
    type: integer
    format: int64

  dateParam:
    name: date
    in: query
    description: Date day
    required: true
    type: string

  incomeParam:
    name: income
    in: query
    description: Expected Income
    required: true
    type: integer



  birthParam:
    name: birth
    in: query
    description: Birth date 
    required: true
    type: string

  nameGoalParam:
    name: name
    in: query
    description: Goal Name
    required: true
    type: string

  valueGoalParam:
    name: value
    in: query
    description: Value number
    required: true
    type: integer

  SavedGoalParam:
    name: saved
    in: query
    description: Saved money until now
    required: true
    type: integer

  dateGoalParam:
    name: date
    in: query
    description: Deadline
    required: true
    type: string

  accountGoalParam:
    name: accountid
    in: query
    description: Account id
    required: true
    type: string

  priorityGoalParam:
    name: priority
    in: query
    description: Priority
    required: true
    type: integer